Additional fixes in benchmarks

+++ bench/bench.h
@@ -1,10 +1,10 @@
 #pragma once
 
 #define START_BENCHMARK(num)			\
-    float __start_time ## num = startBenchmark()
+    double __start_time ## num = startBenchmark()
 
 #define END_BENCHMARK(num)			\
-    float __end_time ## num = endBenchmark()
+    double __end_time ## num = endBenchmark()
 
 #define GET_TIME(num) \
     getTime(__start_time ## num, __end_time ## num)
@@ -12,13 +12,13 @@
 
 typedef void (*BenchmarkFunc)();
 
-float startBenchmark();
-float endBenchmark();
-float getTime(float start, float end);
+double startBenchmark();
+double endBenchmark();
+double getTime(double start, double end);
 
-float performBenchmark(BenchmarkFunc fn);
-float performBenchmarkTimes(BenchmarkFunc fn, int times);
+double performBenchmark(BenchmarkFunc fn);
+double performBenchmarkTimes(BenchmarkFunc fn, int times);
 
-void reportElapsedTime(const char *bench_name, float time);
-void procReportElapsedTime(const char *bench_name, float time);
-void rootReportElapsedTime(const char *bench_name, float time);
+void reportElapsedTime(const char *bench_name, double time);
+void procReportElapsedTime(const char *bench_name, double time);
+void rootReportElapsedTime(const char *bench_name, double time);
